+++ src/overlords/colonization/claimer.ts
@@ -39,7 +39,7 @@ export class ClaimingOverlord extends Overlord {
 				claimer.task = Tasks.claim(this.room.controller!);
 			}
 		} else {
-			// claimer.task = Tasks.goTo(this.pos, {travelToOptions: {ensurePath: true}});
+			// claimer.task = Tasks.goTo(this.pos, {moveOptions: {ensurePath: true}});
 			claimer.goTo(this.pos, {ensurePath: true, preferHighway: true});
 		}
 	}