+++ src/overlords/core/scout.ts
@@ -28,7 +28,7 @@ export class ScoutOverlord extends Overlord {
 
 	private handleScout(scout: Zerg): void {
 		if (!scout.pos.inRangeTo(this.pos, 3)) {
-			// scout.task = Tasks.goTo(this.pos, {travelToOptions: {range: 3}});
+			// scout.task = Tasks.goTo(this.pos, {moveOptions: {range: 3}});
 			scout.goTo(this.pos, {range: 3});
 		}
 	}