+++ src/utilities/utils.ts
@@ -24,8 +24,6 @@ export function getUsername(): string {
 	return 'ERROR: Could not determine username.';
 }
 
-import lodashMinBy from 'lodash.minby';
-
 interface toColumnsOpts {
 	padChar: string,
 	justify: boolean
@@ -78,10 +76,25 @@ export function derefCoords(coordName: string, roomName: string): RoomPosition {
 }
 
 export function minBy<T>(objects: T[], iteratee: ((obj: T) => number)): T {
-	return lodashMinBy(objects, iteratee);
+	let minObj: T;
+	let minVal = Infinity;
+	for (let i in objects) {
+		if (iteratee(objects[i]) < minVal) {
+			minVal = iteratee(objects[i]);
+			minObj = objects[i];
+		}
+	}
+	return minObj!;
 }
 
 export function maxBy<T>(objects: T[], iteratee: ((obj: T) => number)): T {
-	let maxByIteratee: ((obj: T) => number) = (obj => -1 * iteratee(obj));
-	return lodashMinBy(objects, maxByIteratee);
+	let maxObj: T;
+	let maxVal = -Infinity;
+	for (let i in objects) {
+		if (iteratee(objects[i]) > maxVal) {
+			maxVal = iteratee(objects[i]);
+			maxObj = objects[i];
+		}
+	}
+	return maxObj!;
 }